Don't force RTC us_to_ticks to 32-bit
This just doesn't make sense

+++ MicroFrameworkPK_v4_3/DeviceCode/Targets/Native/STM32F10x/DeviceCode/drivers/Timer/netmf_timers.cpp
@@ -641,7 +641,7 @@ UINT64 CPU_MicrosecondsToTicks( UINT64 uSec, UINT16 Timer )
 	}
 	else if(Timer == RTC_32BIT)
 	{
-		return (UINT32)( (UINT64)uSec * (UINT64)timerFrequency / (UINT64)CLOCK_COMMON_FACTOR );
+		return ( (UINT64)uSec * (UINT64)timerFrequency / (UINT64)CLOCK_COMMON_FACTOR );
 	}
 }
